refactor: move new range metrics methods into their old place and remove the deprecation
