+++ aeon/benchmarking/metrics/anomaly_detection/__init__.py
@@ -26,18 +26,20 @@ from aeon.benchmarking.metrics.anomaly_detection._continuous import (
     roc_auc_score,
     rp_rr_auc_score,
 )
+from aeon.benchmarking.metrics.anomaly_detection._range_metrics import (
+    range_f_score,
+    range_precision,
+    range_recall,
+)
+from aeon.benchmarking.metrics.anomaly_detection._range_ts_metrics import (
+    ts_fscore,
+    ts_precision,
+    ts_recall,
+)
 from aeon.benchmarking.metrics.anomaly_detection._vus_metrics import (
     range_pr_auc_score,
     range_pr_roc_auc_support,
     range_pr_vus_score,
     range_roc_auc_score,
     range_roc_vus_score,
-)
-from aeon.benchmarking.metrics.anomaly_detection.range_metrics import (
-    range_f_score,
-    range_precision,
-    range_recall,
-    ts_fscore,
-    ts_precision,
-    ts_recall,
 )
