+++ aeon/benchmarking/metrics/anomaly_detection/_range_metrics.py
@@ -6,19 +6,15 @@ __all__ = ["range_precision", "range_recall", "range_f_score"]
 import warnings
 
 import numpy as np
-from deprecated.sphinx import deprecated
 
-from aeon.benchmarking.metrics.anomaly_detection._util import check_y
-from aeon.utils.validation._dependencies import _check_soft_dependencies
-
-
-# TODO: Remove in v1.2.0
-@deprecated(
-    version="1.1.0",
-    reason="range_precision is deprecated and will be removed in v1.2.0. "
-    "Please use ts_precision from the range_metrics module instead.",
-    category=FutureWarning,
+from aeon.benchmarking.metrics.anomaly_detection._range_ts_metrics import (
+    _binary_to_ranges,
+    _ts_precision,
+    _ts_recall,
 )
+from aeon.benchmarking.metrics.anomaly_detection._util import check_y
+
+
 def range_precision(
     y_true: np.ndarray,
     y_pred: np.ndarray,
@@ -28,13 +24,15 @@ def range_precision(
 ) -> float:
     """Compute the range-based precision metric.
 
-    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_. This
-    implementation uses the community package `prts <https://pypi.org/project/prts/>`_
-    as a soft-dependency.
+    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_.
 
     Range precision is the average precision of each predicted anomaly range. For each
     predicted continuous anomaly range the overlap size, position, and cardinality is
     considered. For more details, please refer to the paper [1]_.
+
+    The `alpha` parameter for the existence reward was removed. Because precision
+    emphasizes prediction quality, there is no need for an existence reward and this
+    value should always be set to 0.
 
     Parameters
     ----------
@@ -43,9 +41,7 @@ def range_precision(
     y_pred : np.ndarray
         Anomaly scores for each point of the time series of shape (n_instances,).
     alpha : float
-        Weight of the existence reward. Because precision by definition emphasizes on
-        prediction quality, there is no need for an existence reward and this value
-        should always be set to 0.
+        DEPRECATED. Default is 0 = no existence reward.
     cardinality : {'reciprocal', 'one', 'udf_gamma'}
         Cardinality type.
     bias : {'flat', 'front', 'middle', 'back'}
@@ -64,10 +60,13 @@ def range_precision(
        1920–30. 2018.
        http://papers.nips.cc/paper/7462-precision-and-recall-for-time-series.pdf.
     """
-    _check_soft_dependencies("prts", obj="range_precision", suppress_import_stdout=True)
-
-    from prts import ts_precision
-
+    if alpha != 0:
+        warnings.warn(
+            "The alpha parameter should not be used in range precision. This "
+            "parameter is removed in 1.3.0.",
+            stacklevel=2,
+            category=FutureWarning,
+        )
     y_true, y_pred = check_y(y_true, y_pred, force_y_pred_continuous=False)
     if np.unique(y_pred).shape[0] == 1:
         warnings.warn(
@@ -75,16 +74,14 @@ def range_precision(
             stacklevel=2,
         )
         return 0.0
-    return ts_precision(y_true, y_pred, alpha=alpha, cardinality=cardinality, bias=bias)
+
+    y_pred_ranges = _binary_to_ranges(y_pred)
+    y_true_ranges = _binary_to_ranges(y_true)
+    return _ts_precision(
+        y_pred_ranges, y_true_ranges, gamma=cardinality, bias_type=bias
+    )
 
 
-# TODO: Remove in v1.2.0
-@deprecated(
-    version="1.1.0",
-    reason="range_recall is deprecated and will be removed in v1.2.0. "
-    "Please use ts_recall from the range_metrics module instead.",
-    category=FutureWarning,
-)
 def range_recall(
     y_true: np.ndarray,
     y_pred: np.ndarray,
@@ -94,9 +91,7 @@ def range_recall(
 ) -> float:
     """Compute the range-based recall metric.
 
-    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_. This
-    implementation uses the community package `prts <https://pypi.org/project/prts/>`_
-    as a soft-dependency.
+    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_.
 
     Range recall is the average recall of each real anomaly range. For each real
     anomaly range the overlap size, position, and cardinality with predicted anomaly
@@ -132,10 +127,6 @@ def range_recall(
        1920–30. 2018.
        http://papers.nips.cc/paper/7462-precision-and-recall-for-time-series.pdf.
     """
-    _check_soft_dependencies("prts", obj="range_recall", suppress_import_stdout=True)
-
-    from prts import ts_recall
-
     y_true, y_pred = check_y(y_true, y_pred, force_y_pred_continuous=False)
     if np.unique(y_pred).shape[0] == 1:
         warnings.warn(
@@ -143,16 +134,14 @@ def range_recall(
             stacklevel=2,
         )
         return 0.0
-    return ts_recall(y_true, y_pred, alpha=alpha, cardinality=cardinality, bias=bias)
+
+    y_pred_ranges = _binary_to_ranges(y_pred)
+    y_true_ranges = _binary_to_ranges(y_true)
+    return _ts_recall(
+        y_pred_ranges, y_true_ranges, alpha=alpha, gamma=cardinality, bias_type=bias
+    )
 
 
-# TODO: Remove in v1.2.0
-@deprecated(
-    version="1.1.0",
-    reason="range_f_score is deprecated and will be removed in v1.2.0. "
-    "Please use ts_fscore from the range_metrics module instead.",
-    category=FutureWarning,
-)
 def range_f_score(
     y_true: np.ndarray,
     y_pred: np.ndarray,
@@ -165,13 +154,14 @@ def range_f_score(
 ) -> float:
     """Compute the F-score using the range-based recall and precision metrics.
 
-    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_. This
-    implementation uses the community package `prts <https://pypi.org/project/prts/>`_
-    as a soft-dependency.
+    Range-based metrics were introduced by Tatbul et al. at NeurIPS 2018 [1]_.
 
     The F-beta score is the weighted harmonic mean of precision and recall, reaching
     its optimal value at 1 and its worst value at 0. This implementation uses the
     range-based precision and range-based recall as basis.
+
+    The `p_alpha` parameter for the potential existance reward in the calculation of
+    range-based precision was removed. `p_alpha` should always be set to 0, anyway.
 
     Parameters
     ----------
@@ -183,8 +173,7 @@ def range_f_score(
         F-score beta determines the weight of recall in the combined score.
         beta < 1 lends more weight to precision, while beta > 1 favors recall.
     p_alpha : float
-        Weight of the existence reward for the range-based precision. For most - when
-        not all - cases, `p_alpha` should be set to 0.
+        DEPRECATED. Default is 0 = no existence reward for precision.
     r_alpha : float
         Weight of the existence reward. If 0: no existence reward, if 1: only
         existence reward.
@@ -208,9 +197,13 @@ def range_f_score(
        1920–30. 2018.
        http://papers.nips.cc/paper/7462-precision-and-recall-for-time-series.pdf.
     """
-    _check_soft_dependencies("prts", obj="range_recall", suppress_import_stdout=True)
-
-    from prts import ts_fscore
+    if p_alpha != 0:
+        warnings.warn(
+            "The p_alpha parameter should not be used. This parameter is removed "
+            "in 1.3.0.",
+            stacklevel=2,
+            category=FutureWarning,
+        )
 
     y_true, y_pred = check_y(y_true, y_pred, force_y_pred_continuous=False)
     if np.unique(y_pred).shape[0] == 1:
@@ -219,13 +212,18 @@ def range_f_score(
             stacklevel=2,
         )
         return 0.0
-    return ts_fscore(
-        y_true,
-        y_pred,
-        beta=beta,
-        p_alpha=p_alpha,
-        r_alpha=r_alpha,
-        cardinality=cardinality,
-        p_bias=p_bias,
-        r_bias=r_bias,
-    )
+
+    y_pred_ranges = _binary_to_ranges(y_pred)
+    y_true_ranges = _binary_to_ranges(y_true)
+
+    precision = _ts_precision(y_pred_ranges, y_true_ranges, cardinality, p_bias)
+    recall = _ts_recall(y_pred_ranges, y_true_ranges, cardinality, r_bias, r_alpha)
+
+    if precision + recall > 0:
+        fscore = ((1 + beta**2) * (precision * recall)) / (
+            beta**2 * (precision + recall)
+        )
+    else:
+        fscore = 0.0
+
+    return fscore
